ARM and AArch64 task and core handling refactor.
The first part of this commit refactors the Task and TaskContext
structures to remove a circular dependency and allow the compiler
to optimize away thread-local mapping on AArch64.

The second part of this commit adds recursive memory mapping for ARM
to make it easier to map thread-local tables for editing.

The third part of this commit refactors how the kernel handles ARM
and AArch64 cores to allow pinning a task to a core when mapping a
page locally. Cores are now added to an array based on the order
they occur in the DTB, and a reverse-lookup map is available to get
a core's index from its affinity value. The Task structure's affinity
tracking now supports a bitmap instead of a single core index.

The fourth part of this commit updates the start code of both ARM
and AArch64 to add a secondary entry point for secondary cores so
that testing MPIDR(_EL1) is not necessary.